Implementing support for secrets
Add configuration to define secrets for specific services for specific apps. These
secrets will be mounted in the container under `/run/secrets/<name>` so that the
service is able to use e.g. database credentials.

Additionally, PREvant starts to use secstr so that credentials are not accidentally
leaked to the container log.