Implement Start and Stop of Services
In order to test the behaviour of services when resources are not available,
PREvant now supports the status change of services. Therefore, PREvant offers
a new resource which allows to start and stop containers. Additionally, the UI
provides buttons, utilizing this resource.

Additionally, this commit cleans up the backend structure and it also uses the
store to reflect the change to the application (cf. #14 for further
information).