Merge default and custom configs when deploying a companion
When a companion is manually deployed, the config is now created by taking the default config for labels, env-vars and volumes defined in the config file and writing the deployment config over it. As a consequence, the default entries are kept if not explicitly overwritten.
Fix https://github.com/aixigo/PREvant/issues/19.