Fix: Apply Templating With Resolved Port Mapping
PREvant did not include the resolved image ports during the templating because
the templating happened before resolving the port mappings. This commit fixes
this issue by using a struct that collects the data and it applies the
templating as the very last step when it will be converted into the list of
service configurations.