Merge PREvant's Traefik Route with Services Routes
When a new service will be deployed to underlying infrastructure,
PREvant will ask the infrastructure which Traefik routing options are
assigned to PREvant itself and if the infrastructure is able to
determine this routing information, PREvant will merge this information
with services routing information.

This feature is currently only implemented for Kubernetes and uses
Kubernetes' downwardAPI. A label file has to be mounted into the
deployemnt (by default it is `/run/podinfo/labels`) and it then queries
the Kubernetes API for all Traefik ingresses that match these labels.

Additionally, the commit improves Docker build caching for having
faster, local development turn around times.