Fix Duplication and Deletion of Apps
Due to the removal of vue-resource in a previous commit (because it does
not seem to be maintained) the duplication and deletion of apps did not
work any more. This commit now moves the functionality to the Vue store
and uses native fetch API to perform these operations. Additionally,
this improve the user experience a little because it gets rid of page
refreshes.