Persistence Hooks for Storage Path Management
This commit will allow users to share declared volumes and delete volume
paths exposed by the docker manifest with the help of Persistence Hooks.
The volume paths are processed according to the javascript functions
and returned to PREvant in an array of array, where each element of the
array is pair of service name and volume path. This will then be passed
to the infrastructure for storage creation.

The hooks also allows adding of any additional volume paths that are
undeclared in the image.