Improve Kubernetes Namespace Handling
Before this commit it wasn't possible to create applications with names
like MY-APP on a Kubernetes cluster because the application name has
been used as namespace identifier but the name could contains characters
that weren't allowed in RFC1123.

Futher, this commit adds an additional configuration option to add
optional annotations to namespaces. For example, with
"field.cattle.io/projectId" it is possible to link dynamically created
namespaces to Rancher projects.

Also, the number of API calls to the Kubernetes backend has been reduced
by just deleting the namespace of the application rather then deleting
all objects in the namespace first.

Additionally, this commit updates a set of library dependencies.