Improve Jira Handling
Add apiKey option in order to use API Keys instead of username and
password.

Additionally, an empty dashboard with active Jira configuration resulted
in an HTTP 500 error because the Jira query failed on the Jira API
level. If no applications are available, there is no need to query the
Jira API.