Add Support for Multi-Platform Image Manifests
This commit add support for retrieving image manifest if the images
provides multiple platform variants, e.g. amd64, arm64, …. It also
provide better HTTP responses if the API client provided an image that
doesn't exist, if the authentication for the image doesn't work or if
anything else failed for the image manifest retrieval.