Fix Duplicate Deployments in Bootstrapped Apps
If the application is bootstrapped by a bootstap container and then
companions have been converted into instances, and then the instances
should be updated with a subsequent HTTP request, the application ends
up with a second version of the instance. This circumstance is not
intended and this commit makes sure that subsequent HTTP requests update
existing deployments while taking the bootstrapping into account.