Make Kubernetes More Robust
In some cases it is necessary to access the Kubernetes namespace ID in
bootstrap container arguments. For example, some Helm charts require
that you provide --namespace option if a bootstrap container utilizes
Helm under the hood. This commit provides {{infrastructure.namespace}}
template argument for bootstrap container start up

There was also a bug that conversion of ingress routes included multiple
stripPrefix middlewares when derived from bootstrap containers. The once
provided by the bootstrap container must not included the defaults from
PREvant and thus have to filtered.

Additionally, there was a bug in the Kubernetes middleware ID
generation. Capital letters in the application names caused the
deployment to fail because these ID have to be RFC1123 compliant. This
commit ensures that names will be generated according to this RFC.