Improve App Name Handling
This commit makes sure that app names will be limited to valid
characters for Kubernetes backend. Also the duplicate UI shows the
client error now.

Additionally, the filtering of app names in the dashboard will be
persisted in the URL route so you can refresh the pages and keep the app
name filtering.