Prevent endless updating when a service is paused
In case of a paused app, the Kubernetes backend endlessly causes the web
host meta cash to crawl web host meta information. Thus, it floods all
browser to refresh all the time which then increases the load on the
server.

This commit fixes makes the started_at value optional, so that now
Utc::now usage is necessary.