Configure Static Web Host Meta
For services that are outside of the control of the team that uses
PREvant and where is no possibility to create a resource route for
.well-known/host-meta.json it is now possible to configure a static URL
to the OpenAPI documentation that will be then integrated into the
PREvant dashboard. This file will be massaged so that it points to the
service service URL.

Also, this commit makes sure that Kubernetes Ingresses with path prefix
will be mapped correctly. Without this you won't be able to expose
multiple prefixes in a bootstrapped environment.

Additionally, PREVant provide a new CLI option to debug new config from
the outside and reduces some network calls.