Integration test with Kubernetes and Docker
This commit provide integration testing for Kubernetes and Docker. For
each PR the Docker image will be built and deployed via Testcontainers
and K3d and then the deployed images will be tested in both
environments.

Also, this commit provides a fix in the URL detection when the host or
forwarded host contains already the port.