Hook for Mapping ID Token Claims
In some situations the ID token, used by authentication, is generated on
behalf of someone and thus, the sub field doesn't match to the user's
sub field on the PREvant dashboard. For example, in GitLab CI tokens can
contain the user_id but not in the sub field.

In order to be able to have this mapping, there is a new hook that
allows transforming the ID token claims to the corrected owner.