Persistent Worker-Producer Queue
This commit provides the option to operate PREvant with a persistent
database connection. This configuration provides the ability to preserve
the creation and deletion of applications beyond restarts of PREvant.
Also, this gives the advantage to have multiple instances of PREvant
running at the same time.

Also, the database feature provides the foundation to implement a solid
back up and restore functionality (see #149).

Fixes #40