Cache the Ticketing Information
In order to reduce the HTTP traffic to the ticketing system, i.e. Jira,
PREvant crawls the ticketing information in the background and caches
the responses (similarly to the host meta caching). Comparing the
Prometheus metrics for a day (without and with the change), we can see
that the network traffic averages below 1.0 kp/s where before it varied
between 2.0 kp/s and 12.5 kp/s (checkout the PR description to see the
graphs).