Reduce Memory Consumption
In deployments with a large set of services the memory consumption of
PREvant increases a lot because with the number of services the entries
in FuturesUnordered increases. This increase of entries will consume a
lot of memory.

Thus, with this commit the crawling of host-meta entries will be chunked
by application name, so that less memory will be consumed.