Do not wait until the dawn of time for bootstrapping
When polling the status of a bootstrapping pod, PREvant polled the
status in an endless loop. This approach can cause an endless loop due
to the fact that an invalid image reference for the bootstrapping
container does not result in a specific error phase, it rather leads to
an endless pending status. Therefore, the polling of the bootstrapping
pod status is limited to one minute.