Improve ux for required auth (#273)
When PREvant is configured to require auth then the user needs to be logged in
in order to perform writing operations such as creating tiles or deleting
ones. However, there was no feedback in the UI if the user has to be logged
in to perform this action. With this change the duplication/deletion dialogs
can still be opened but the actual inputs to perform the action are disabled
and a warning alert is displayed.