Avoid Flickering in Frontend
This changes fixes a race condition in the optimistic UI updates: when
an application has been deleted or duplicated, it could be that the
optimistic UI got overwritten by the applications server side event
stream and the latest update from the server didn't contain the deletion
or duplication yet.

Additionally, this change upgrade frontend dependencies and integration
the Vue dev tools for better debugging.