Merge back-up and restore tasks
This change makes sure that the back-up and restore tasks are merged if
possible. If they cannot be merged, they will be executed until without
merging to make sure that the application state is the same as if they
are executed in singular tasks.

Also, this refactoring fixes #277 for queue processing without
persistent database queue.